Added a general mechanism for regression testing to ensure that Chapel code is executed in parallel, added a test to ensure that whole array assignment is parallel, and fixed a bug related to passing line number and filename information through cobegin and coforall statements.
General Mechanism for Regression Testing of Parallelism
-------------------------------------------------------

This is controlled by a config param called chpl__testParFlag, which
is false by default.  To test a section of code, turn the verbose
output on by calling chpl__testParStart.  Turn the verbose output off
by calling chpl__testParStop.

The function chpl__testPar(...) outputs a message, currently when a
default arithmetic domain follower is invoked.

Testing that Whole Array Assignment is Parallel
-----------------------------------------------

Added a test to ensure that whole array assignment is parallel.
Basically:

  chpl__testParStart();
  A = B;
  chpl__testParStop();

With

  COMPOPTS: -schpl__testParFlag=true
  EXECOPTS: --maxDataParallelism=2 --limitDataParallelism=false

And a PREDIFF that sorts the verbose parallel testing output (lines
starting with CHPL TEST) in place.

Bug Fix
-------

Fixed a bug in which line numbers and filenames were not passed
through coforall or cobegin statements, resulting in garbage error
messages due to uninitialized accesses to line numbers and filenames.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16913 3a8e244f-b0f2-452b-bcba-4c88e055c3ca