Compiler --
compiler/passes/insertTransactions.cpp:
Added couple of new primitives: PRIM_TX_CHPL_ALLOC, 
PRIM_TX_CHPL_ALLOC_PERMIT_ZERO, and PRIM_TX_CHPL_FREE which are 
STM equivalents of PRIM_CHPL_ALLOC, RIM_CHPL_ALLOC_PERMIT_ZERO, 
and PRIM_CHPL_FREE respectively. 

In the case of PRIM_TX_CHPL_ALLOC* primitives, we leave the 
PRIM_MOVE part of the call expression as is. In that sense, these 
primitives are handled differently from the rest of the stm 
specific primitives. Ideally, I need to insert a temp that captures 
what ever is returned by the malloc and then assign this temp to the 
target of the PRIM_MOVE. 

In the case of PRIM_TX_CHPL_ALLOC* primitives, we also create a new 
description by adding the prefix "stm " to the existing description 
name. These new descriptions are codegen'ed similar to their 
non-stm equivalents and are added to chpl_memDescs at compile time. 

compiler/AST/primitive.cpp:
As far as I could tell, there was no direct way of getting the 
description name from the chpl_alloc (or chpl_alloc_permit_zero) call 
expression. Therefore, I added a new map data-structure when given a 
VarSymbol* gives the description string associated with it. The 
newMemDesc function (in primitive.cpp), is responsible for populating 
this map data-structure in addition to its original operation which 
was to create a VarSymbol* given a description string.

compiler/AST/expr.cpp:
For the most part, codegen handles these primitives similar to their 
non-stm counterparts, especially given the fact that we do not alter 
the PRIM_MOVE part of the call expression. 

In the case of PRIM_TX_CHPL_FREE, we do not explicitly set the memory 
pointer to NULL because we cannot actually free the memory until the 
transaction is guaranteed to commit. 


Runtime -- 

chpl_mem.h: 
We have be using the following naming scheme -- the enum constants used 
by the runtime stm layer starts with "CHPL_RT_MD_STM_" instead of 
"CHPL_RT_MD". The corresponding descriptions begin with "stm". Added a 
pair of enum constants / descriptions to represent the STM equivalents
 of CHPL_RT_MD_CHAPEL_CODE and CHPL_RT_MD_ARRAY_ELEMENTS. 

The rest of the runtime modifications are to accommodate the new 
compiler primitives mentioned above. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@16928 3a8e244f-b0f2-452b-bcba-4c88e055c3ca