Added a test to ensure that the STREAM benchmark is run with 2 tasks per locale when using the execute options:
  --maxDataParallelism=2 --limitDataParallelism=false

This uses the same mechanism as the other parallel test cases in
arrays/deitz/parallelism.  Added output to the Block distribution when
compiling with the compiler option "-schpl__testParFlag=true".

Unfortunately, removing --limitDataParallelism=false from the list of
execute options makes this test fail.  We seem to be off by one in the
Block distribution when running on one locale.  On more than one
locale, the number of tasks we use seems to be non-deterministic.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16941 3a8e244f-b0f2-452b-bcba-4c88e055c3ca