Implemented sparse domain assignment and index removal
This commit implements assignment for sparse domains and in doing so
changes the way we do sparse domain assignment for associative domains
as well.  It also adds support for removing indices from a sparse
domain which we'd never bothered to implement before.  Finally it
fixes a subtle bug in associative domain index removal that we had not
run into before.

The main change in this commit is that we used to use the
getIndices()/setIndices() calls for assignment of sparse and
associative domains, but these were never a particularly good match
since these domain types are more like arrays (in that they use
O(numIndices) storage rather than O(1)).  As a result of the
interface, domain assignment was done in a fairly hacky and fragile
way for associative domains and was never implemented for sparse
domains.

As a result of this change, we only use the setIndices/getIndices
calls when both domains are arithmetic.  If they are not, then we use
iteration to do the index set assignment.  In particular, we first
iterate over all of the LHS domain's indices, removing any that aren't
in the RHS domain; then we iterate over all of the RHS domain indices,
adding any that aren't in the LHS domain.  This was done rather than
the conceptually simpler method of clearing the LHS domain and adding
the indices from the RHS domain in order to preserve array elements in
the intersection of the two domains without making a complete backup
of the arrays (as we used to do for associative domains).

This assignment is currently serial, though it should eventually be
parallel.  Added a warning to this effect to match the other warnings
that Steve's been adding recently.

A nice thing about this change is that it doesn't require any new
methods on our sparse/associative distributions since we're just using
the normal iterator and add/remove concepts.  We had originally
envisioned needing to add a getIndices iterator for these domain
types.

Used this logic for domain initializations as well, but used the
simpler form of only adding indices since we can rely on the LHS to be
empty to start.

This change served as a forcing function to finally implement index
removal for sparse domains so I did that for our default sparse
distribution.  I still need to come back and do it for our
(undocumented) CSR distribution but that will be a separate check-in.
This required polluting the abstract array classes with a new method.
Steve and I discussed an approach of avoiding this pollution by using
dynamic casts rather than dynamic dispatch, but this would require us
to move eltType to BaseArray -- arguably a good move, but one that we
can't do today due to our poor support for subclassing generic
classes.  This would potentially be a good summer project for
Jonathan.

Removed the dsiGetIndices, dsiSetIndices, and dsiReallocate calls from
the sparse and associative distributions since they're no longer
required.

Finally, this change found a bug in our DefaultAssociative
distribution in which we failed to reset a lazily-deleted array
element to the default value, so if a subsequent add reclaimed that
spot, it would inherit the old value.  Added this logic to the
dsiRemove() call since it's very dependent on the data structure in
question (e.g., our current sparse data structures do not require a
similar reset because they actively reclaim the space).

Updated the STATUS file to reflect the new support for remove and
assignment on sparse domains and removed two futures for domain
assignment.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16945 3a8e244f-b0f2-452b-bcba-4c88e055c3ca