Added three futures that simplify the future submitted by John regarding a distributed array of sync ints.  The test test_block_array_of_syncs creates a Block-distributed array of sync ints and fails when initializing it with the value -1.  The test test_block_array_of_ints_assigned_class_field declares an array of ints and fails when looping over the array with a forall and assigning each element a field in a class.  The test test_iterator_with_local_around_yield declares an iterator that yields within a local block.  If this iterator is inlined, then the local block ends up being applied to the entire body of the loop.  This is the root of the problem. Followers with local blocks around yield statements are getting inlined, but the body of the loop requires some communication.  For the array of sync case, that communication is quite unfortunate.  The code in question is
  S$ = -1;

where S$ is an array of sync ints.  The whole array assignment is done with the
forall.  The rhs argument to the assignment function is -1.  In the
implementation of the forall which inlines a leader with an on-statement, the
reference to the argument -1 is not remote value forwarded due to sync writes
within the body of the loop.  So then the remote references violate the local
block.

Updated all of these futures to match.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16947 3a8e244f-b0f2-452b-bcba-4c88e055c3ca