This change adds a launcher for the XMT.
This change adds a launcher for the XMT. It has three options: --numNodes,
or -nn, which sets the max_procs and num_procs values equally; --numTeams,
which sets the num_procs value; and --maxProcs, which sets the max_procs
value. These get passed to mtarun. Note that this change also fixes a
build error whereby the archiver, ar, was defaulted to the XMT version,
which doesn't work with gcc on the launcher build. The Makefile change
forces the ar to use /usr/bin/ar.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16951 3a8e244f-b0f2-452b-bcba-4c88e055c3ca