Fleshing out sparse domain/array iteration a bit
Added support for a serial implementation of leader/follower for
sparse arrays which we've never supported before.  This generates
a warning like other serial leader/follower iterators.

Improved the thing passed from the leader to the follower from a
bool (which was a completely bogus placeholder) to "this".  The
follower can then make sure that "follower" matches its "this" 
to ensure that it isn't being zippered with anything (other than
optionally itself).  This improves our ability to only zipper
things together that can deal with it (until we better define
zippered iteration for leader/follower or support standalone
parallel iterators).

Updated the warning for the release sparse tutorial example which
changed a little due to this change.  Added a test of parallel sparse
domain and array iteration which did not work previously.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16968 3a8e244f-b0f2-452b-bcba-4c88e055c3ca