Added diagnostic to indicate when the fast on clause optimization has been applied.
When the --report-optimized-on flag is used during compilation, a
message is printed about on clauses that were optimized for potential
fast remote fork.  When not in developer mode, the report only
includes on clauses in the user's code (not internal or other Chapel
modules).

Moved the fast on clause optimization to its own file in
compiler/optimizations/optimizeOnClauses.cpp (and removed it from
passes/parallel.cpp).

Added a some basic tests that will make sure the optimization is
working.  In addition, added a noexec test for RA
(test_OptimizedOn_RA.chpl) that makes sure the optimization has been
applied to the main timed loop of RA.  This test uses the ra module
that is defined in release/examples/hpcc/ra.chpl, so the result of
this test directly reflects our release version of RA.  If ra.chpl (or
any of the modules it uses) changes, there may need to be changes in
the .good output for the test.  Also, if the directory structure
changes, then the compopts file needs to be updated with the correct
module include path.

Realphabetized the Misc developer flags since I was in there anyways.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16969 3a8e244f-b0f2-452b-bcba-4c88e055c3ca