Fixed a critical bug in which a local block would be incorrectly applied to the body of a loop when the compiler inlines an iterator that has a yield statement in a local block.  This made it so that the body of a forall loop over a single Block-distributed array could not contain any communication.
This fix involves the introduction of an "unlocal block" marked by the
primitive PRIM_BLOCK_UNLOCAL.  During the replacement of yield
statements when inlining iterators, if the yield statement is in a
local block, then the compiler will insert an unlocal block around the
copied body of the loop (or a call to the body of the loop).

After inlining iterators, during the fragmentation of local blocks
(see r14785), unlocal blocks will be removed.

Fragmentation of local blocks could not be moved before iterator
inlining because then the local block wouldn't apply to code in the
iterators that are inlined (see r16240).

Removed 5 futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16971 3a8e244f-b0f2-452b-bcba-4c88e055c3ca