Added support for assigning one distribution to another if they are of the same type and there are no domains declared over the distribution on the left-hand side of the assignment.  If there are domains declared over this distribution, then we will need to redistribute these domains (but this support is not yet implemented).
This support requires a new method called dsiAssign that updates
'this' (the left-hand distribution object) with its argument (the
right-hand distribution object).  For the default distribution, the
dsiAssign method is empty since there is no state associated with the
default distribution.

Implemented dsiAssign for Block and Cyclic.  Added
Block.dsiReprivatize to support reprivatization on Block
distributions.  Changed the default index type of Cyclic distributions
to int (from int(64)).

Changed the fields in Block and Cyclic to variables since they are now
changed outside of the constructor.

Added special handling for initializing variables of distribution
types in the compiler since these need to be potentially privatized
values.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16972 3a8e244f-b0f2-452b-bcba-4c88e055c3ca