On second thought, broke the modules/dmaps directory into two subdirectories: layouts: intended for use to define classes that characterize single-locale domain maps, and dists: intended for use to define classes that characterize multi-locale domain maps.
As part of this change, I also removed the dmapModPath variable from
util/files.cpp (formerly distModPath I think) because the intention
was for this directory to be included in the standard module path
since the standard distributions and layouts are standard modules.

Rationale for this change: Once I saw it, having the directory name
'dmap' concerned me both because it used the term for something that
is not the value portion of a domain map (i.e., "Block" is not a dmap,
it's a class used to characterize a dmap) and because it was a
non-keyword-based use of dmap which seems likely to increase the
chances that we use the term as a noun in conversation (rather than
"domain map").

I think breaking the directory into two also has the benefit that as
the number of distributions and layouts grows, it's a natural
distinction between the two categories and will help avoid a single
directory growing too large.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16990 3a8e244f-b0f2-452b-bcba-4c88e055c3ca