Fixed bug in associative domain assignment
This commit fixes a bug in the associative domain's dsiIndsIterSafeForRemoving()
method in which if enough indices were removed, the domain would be resized
and the iteration and domain would be corrupted.  Added a variable, postponeResize,
that can be used to skip resize steps for a period of time and used this in
the iterator.  Then do the resize at the iterator if applicable.

Added a test that didn't work prior to this change, but does work now.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16991 3a8e244f-b0f2-452b-bcba-4c88e055c3ca