Committing David's Cyclic constructor change on his behalf
This commit changes the order of the arguments to the Cyclic
constructor such that the rank and idxType are inferred from the
starting index to "deal indices" from, now named startIdx.

Other changes are to tests that invoke the Cyclic constructor to
simplify them and remove the redundant arguments.

David was finding that this commit was making robustness test
basics/test_zipper_default.chpl to fail with a "zippered iterations
have non-equal lengths" error, but from what I could tell that error
was occurring prior to this commit as well, and the changes to
CyclicDist themselves look very clean, so I consider this to be a step
forward and orthogonal to the bug.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16992 3a8e244f-b0f2-452b-bcba-4c88e055c3ca