Fixing intentional regression with zippered sparse iteration
Last week, in cleaning up the sparse domain/array code I changed the
leader iterator to pass "this" to the follower rather than the
pointless placeholder bool that we used to pass.  The idea was that a
follower could then see if the "this" that it was being asked to
follow shared the same sparse domain that it used and to generate a
runtime warning if not (and a compile-time warning if it was being
asked to follow anything other than a sparse domain/array reference).

However, before committing, I only got as far as checking to see
whether the thing we were following was equal to this (i.e., whether
we were following ourself) rather than worrying about the case of
iterating over a sparse domain and its array or two sparse arrays
sharing a domain or ...  My guess was that nobody relied on this and
my hope was to get the fix checked in before the code freeze.

I didn't get the commit in before the code freeze and it turned out
that John Lewis did rely on it, so this is the fix.  I checked this
against the whole test system serially and against the
release/examples and sparse/ test subdirectories under GASNet.

In this new fix:
- Sparse domains will follow themselves or an array declared over
  them.
- Sparse arrays will follow their domain or any array that shares
  their domain.

Going forward we're going to need to figure out what happens if two
sparse domains/arrays with different domains are zippered, what it
means to zipper a sparse domain/array with something that's not
sparse, etc.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17001 3a8e244f-b0f2-452b-bcba-4c88e055c3ca