This changes the amount of default optimization done for CHPL_COMM == gasnet
David pointed out the other week that compilation was taking so long
due to the fact that we were using -O3 on our generated code.  Steve
and I thought that we only used -O3 if -O was used (or --fast which
implies -O).  It turns out that because we were using the recommended
GASNet Makefile variables, we were getting a bunch of GASNet-preferred
optimization options on every compile for CHPL_COMM==gasnet.

This change breaks up our use of the GASNET_CFLAGS variable into its
three component variables: GASNET_MISC_CFLAGS and MANUAL_CFLAGS
(always used) and GASNET_OPT_CFLAGS) (only used when the -O flag is
thrown).  This speeds up the compilation of a test like
test_zipper_default.chpl from ~5.5 minutes on chap01 to ~1 minute when
-O was not specified and should generally speed up testing and
timeouts (which is not to say that we shouldn't still work on
tightening up our generated code to make things even better... :).

Steve asked what would happen if you were using the debug version of
GASNet and the -O flag.  The answer is that you'd get the
GASNET_OPT_CFLAGS flags, but that they wouldn't contain things like
-O3, but we would throw -O3 ourselves, so you'd still get
optimizations.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17012 3a8e244f-b0f2-452b-bcba-4c88e055c3ca