Moved the old taskParallel.chpl example in the primers directory to parallel/taskPar/figueroa to preserve it.  It will be replaced in a later commit with another example.
Committed the move before the new example, because it would likely
cause svn confusion if I moved it out of the way and then added back a
file with the same name in the same commit.

Temporarily removed taskParallel from the Makefile (in case anyone
does any release testing), but I left it in the README, since it won't
be changing with the new example.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17013 3a8e244f-b0f2-452b-bcba-4c88e055c3ca