Fix for an infinite loop in abnormal Chapel exit
This commit fixes an infinite loop that occurs when we fail to join
against the polling thread on a Chapel exit.  The behavior prior to
this commit was to call chpl_internal_error() which itself calls
back into the same exit code and attempts to re-joing the thread.
The new behavior is to use a static variable to detect this loopback
case, avoid joining the thread and call gasnet_exit() to quit the
program.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17016 3a8e244f-b0f2-452b-bcba-4c88e055c3ca