Adding a future showing why we get the "failure to join polling thread" error
This test shows that we get the "failure to join polling thread" error
when multiple threads within a locale hit a halt simultaneously.  I
believe we should really only process one abnormal exit within a
locale and have others spin until the program is torn down.  I have a
change that does this (which also removes the failure to join polling
thread error), but in the interest of respecting the code freeze will
hold off on committing it until after the release unless there's a
groundswell of opinion to get it tested and committed today.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca