Caught CHANGES up to the current state of things
This brings the CHANGES file up to the present in terms of commits.
More needs to be done in terms of wordsmithing, ordering,
categorization, etc.  However, this would be a reasonable time to read
and see if you think the work you've done since the last release is
captured here if you want to.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17026 3a8e244f-b0f2-452b-bcba-4c88e055c3ca