In my earlier commit today, I made it so that the BUILD_VERSION file would only get generated for the developer, thinking that (a) we'd always build with CHPL_DEVELOPER set and (b) a user of the release would have BUILD_VERSION supplied for them by the gen_release script. These assumptions ignore the fact that we don't always build as developers (and moreover, an external user downloading the tree from sourceforge wouldn't necessarily), and most importantly, that our nightly regression tests don't always run as developer.
The fix for this is actually quite simple, which is to have the
Makefile put in a dummy version number (-999) when developer is
not set.  This is actually what we did in the script that I
removed from the release today anyway, so it just puts us back
to where we were with simpler logic and no dependence on a script.

My apologies for the mess.  I thought I'd tested this sufficiently
by running the nightly and testRelease scripts, but I always run
these with CHPL_DEVELOPER set apparently while others' cron jobs
don't.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17034 3a8e244f-b0f2-452b-bcba-4c88e055c3ca