Teasing "Domains and Arrays" chapter into two distinct chapters
This is the start of my spec edits.  We discussed splitting this into
two chapters in our last review of it, and I continue to think it's a
good idea, particularly given the other splits we've been doing.

This split is really, really rough.  I essentially made two copies of
the old chapter, pulled sections from each copy that referred to the
wrong concept (e.g., arrays in domains), removed duplicate labels,
updated references, and updated spec.tex and Organization.tex.  Now
the real work begins.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17038 3a8e244f-b0f2-452b-bcba-4c88e055c3ca