Added CONTRIBUTORS.devel
Added a CONTRIBUTORS.devel file that's intended to capture people
who've contributed to the SVN code base, but whose code does not end
up in the release, for those who work from the SVN tree.  At present,
this list includes Srini and the PNNL guys who helped with the ARMCI
port back in the day (who had previously been listed in CONTRIBUTORS,
which seemed incorrect on review).

Added mention of this file to README.devel.

Also added Srini to the bottom section of CONTRIBUTORS to reflect his
role in helping improve the code quality this past six months.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17058 3a8e244f-b0f2-452b-bcba-4c88e055c3ca