Update the Chapel quick reference card
Added "new", "dmapped", "sync", and "single" to the expression precedence
table and combined the rows for if, forall, [, for, sync, and single.

Added a block domain map to the Domains and Arrays section and dmapped
the arithmetic domain that was already there with it.

Changed: var SD: sparse subdomain(1);
to:      var SD: sparse subdomain(D);

Played with spacing to make it so no subsections are split across columns.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17064 3a8e244f-b0f2-452b-bcba-4c88e055c3ca