Adding a bug submitted by Wei Li
This is an interesting bug.  Compiling the code results in:

typefnbug.chpl:1: In function 'fun':
typefnbug.chpl:1: error: illegal cast from real(32) to real

From the .future:

I'm not quite sure what's going wrong in this test, but my
best guess would be that the logic we use to unify inferred
return types might be trying to be applied to type functions
as well.  That's just a wild guess, though.

Also of interest, the small rewrite captured in typefnbug-works.chpl
(adding an else clause) causes this to work.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17092 3a8e244f-b0f2-452b-bcba-4c88e055c3ca