Filing working test and error message future against one-tuple detupling
The onetuple-error.chpl test shows that in the declaration:

    var (x): (int);

the compiler tries to index into 'int' as though it's a tuple type (in
our current language definition it is not) rather than issuing a
useful error message to the user.  See the .future for more details.

onetuple.chpl shows the correct behavior when using the type 1*int.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17094 3a8e244f-b0f2-452b-bcba-4c88e055c3ca