Restoring runtime/src/stm/gtm/comm/gasnet/Makefile.share back to r17105. The checks that were removed in r17105 shouldn't have been removed in the first place! They are necessary to determine the path of the gasnet.h header file. Duh!
Removing parts of the madness code that is not relevant to stm
modifications. In the long run, maintaining the remaining parts
should also become necessary once the stm support becomes stable
enough to deal with minor changes to madness code.

RA with --ignore-atomic works fine. Working on a fix for the 
"Arithmetic Exception" error when running RA with gtm.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17108 3a8e244f-b0f2-452b-bcba-4c88e055c3ca