Added an optimization to remove empty records from the IR.  Any record type with no fields is removed from the IR.  Variables and arguments of this type are also removed.  The rest of the IR is updated.
Added a 'developer' flag --no-remove-empty-records to disable this
optimization.  It is on by default and it is not meant for the user to
disable (since the primary purpose of the optimization is to eliminate
warnings).  It doesn't do anything deep; the flag is for debug
purposes.

Removed code that added a field to empty iterator records.

This removes a compiler warning "'x' is used uninitialized" that shows
up with --devel and -O and CHPL_COMM=gasnet on ra.chpl.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17109 3a8e244f-b0f2-452b-bcba-4c88e055c3ca