Removed "var" from the array slice methods named "this".
These methods return array records with fields that can be implicitly assigned,
but the record itself cannot be assigned.

This change masks the problem in test/users/jglewis/slice_bug and
test/functions/diten/test_promote_slice because they no longer pass an array
returned from a var function directly as an argument to a promoted function.
Remove the .future files for these tests.

Commented out a line in the slice_bug code that accesses an array out of
bounds.

Added a test that still runs into the same problem that passes an array
returned from a var function to a function that is defined over a scalar value.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca