Removing square brackets from loops that could iterate over ranges
Javalearner's email that came in today made me wonder (as I often do)
"What are we doing that causes people to write their for loops as:

      for i in [1..n]

rather than

      for i in 1..n

I checked the release/examples directory and found that the
taskParallel case did this, so simplified it.  This is a new test, so
doesn't explain earlier users who've done this, but could explain
where javalearner got it form.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17135 3a8e244f-b0f2-452b-bcba-4c88e055c3ca