Improvements to range striding and slicing
This change was motivated by my (re-)discovery of the fact that we
store range strides using an 'int' even if the index type is an
int(64), implying that we couldn't create a range like:

         1..max(int(64)) by (max(int(64))/2)

even though this seems like a reasonable thing to do.  We also
discussed improving this situation in a code review some time ago, but
it fell by the wayside.  In making this change, I also made some
improvements/fixes to range slicing.

I changed the range's stride from 'int' to the signed int type of the
same size as the range's eltType (I really want to rename this
idxType).  Thus the stride of an int or uint range will be int, the
stride of an int(64) or uint(64) range will be int(64), the stride of
an int(8) or uint(8) range will be int(8), etc.  I implemented this
using a type function defined in ChapelBase.chpl -- typeToSignedType()
-- that takes a type and returns the signed version of that type.  For
anything other than ints it currently just returns the type itself.

[Note that I considered adding another type field to the range called
strideType for full flexibility, but in addition to changing the
range's type signature in some unsavory ways, this also didn't seem to
add much power.  For example, having a stride type of uint(64) rather
than int(64) is not all that helpful since once you're in the range of
values that can be represented by uint(64) and not int(64), you're
only getting 1-2 elements anyway...]

I made a similar change to DefaultArithmetic.chpl because it stores
stride values in the array and they were also hard-coded to be ints.

I changed the 'by' function to take a stride of 'integral' rather than
'int' type and cast this value to the argument range's stride type
before multiplying it by the old stride.

I made the range this (slice/intersection) function return a range of
the same eltType as 'this' rather than the result as it was before.
Under the semantics of slicing, this seems to make more sense (if I
slice into an int range with a uint(64) slice, it seems most logical
to make the result be an int range.  This seems to reverse what we did
before, so hopefully I'm not forgetting something important -- this
seems most logical to me today and didn't appear to break anything.

I extended the extendedEuclid implementation to work on 64-bit ints by
creating a helper function that actually implemented the algorithm and
wrapping it in wrapper functions that expected pairs of ints and pairs
of int(64)s respectively.  There may be a cleaner/better way to do
this.

I did some extra casting to more cleanly permit ranges of different
element types to be intermixed.

I added two tests:

  bigStride.chpl : creates the big range described above to make sure
                   it works (previously it would get a stride of -1)

  mixSliceType.chpl : slices an int(32) range with an int(64) range
                      and vice-versa and makes sure the types are as
                      expected.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17137 3a8e244f-b0f2-452b-bcba-4c88e055c3ca