Adding four user futures
These are the futures I mentioned in today's meeting that relate to
domains with negative strides.  Three of them are from the user,
which I rewrote to avoid using our old syntactic sugar:

      var A: [i in D] = f(i);

I'm not 100% positive what we should do about them -- i.e., what
effect if any should a negative stride have on an array, particularly
a negative stride of -1.  One argument is that since the domain is a
set, striding by -1 doesn't change the set and you get the usual
thing.  Another argument would be that it doesn't change the set but
it does change the default serial iteration order of the set.  The one
thing I'm pretty sure of is that it shouldn't change the memory layout
order of the array's indices (though I suppose this is a decision that
could be made by the distribution author, so I guess I'm only speaking
for the default and standard distributions).

I suspect that the semantic problem is the big problem here, whereas
getting any of the semantics to work right shouldn't be so hard.  So
maybe I should have filed these as semantic futures rather than bugs?

The fourth future uses the old syntactic sugar in a way that ought to
be legal but pointless and instead results in a compiler error (and a
fairly strange one to my mind).  I haven't investigated at all.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca