Making start_test only generate spec tests on top-level invocation
When executed in a test subdirectory (e.g., test/release/examples),
the behavior of start_test has been to only test that directory
downwards.  In such cases it doesn't make sense to generate the spec
tests (and in fact, this causes start_test to fail when run on the 1.1
release unfortunately, which wasn't caught due to changes in my local
copy of gen_release/testRelease).

With this change, we only generate the spec tests if start_test is
executed with no arguments and from $CHPL_HOME or $CHPL_HOME/test.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17144 3a8e244f-b0f2-452b-bcba-4c88e055c3ca