Small improvement to 'illegal use of a function that does not return a value error message
Previously, this error message didn't indicate which function it was
that was being used illegally which could be confusing in a statement
with a lot of function calls.  With this change, I put a function name
into the error message if one could be determined with minimal fuss.
Updated existing tests that triggered this error message.

I retired John Lewis' recent future as a result of this change.  I
thought that he was failing to get a 'syntax error' error message when
his code did not contain a ':' and couldn't understand why it
wouldn't.  With the colon, it makes sense that there would be no
syntax error since the code expresses a cast.  Nevertheless, we were
both concerned that there was some other (compiler introduced?)
function that the compiler was complaining about which is what
motivated this change.  I'll discuss further with John.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17157 3a8e244f-b0f2-452b-bcba-4c88e055c3ca