Fixing incorrect path in testRelease
Somehow I missed this when I changed the structure of the tarball
last week -- the grep which is executed at the end to generate
the email results of testing the release was using the wrong path,
which accounted for the empty release test mails that showed up over
the weekend.  This is the fix which ran last night, producing the
correct output.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17162 3a8e244f-b0f2-452b-bcba-4c88e055c3ca