Added the start of a pass to narrow wide classes and references.
Added a case to code generation not previously exercised in which the
'super' field of a class is accessed and the result needs to be wide
even though the class has been narrowed.

This narrows about 25% of wide references in programs I examined.  It
results in a compilation speedup of approximately 10%.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17164 3a8e244f-b0f2-452b-bcba-4c88e055c3ca