Continued work on the optimization to narrow wide classes and references.
Extended widening of class instances even if the instance is used with
PRIM_SETCID, which supports both narrow and wide class instances.

Removed the call to narrowWideReferences from the passlist and pushed
this call to the end of insertWideReferences to allow some clean-up to
be done after the pass.

Noticed a function called narrowReferences that I added in revision
14803 to do some narrowing of references based on narrowing of
references in local blocks that was particularly important to HPCC
stream.  Since this is now handled in this new optimization, I removed
this old code.

Added code to break a single 'move' into two 'moves' if the left-hand
side is a wide reference of a wide class and the right-hand side is a
class.  Such an idiom was never encountered before.

This increases the effectiveness of narrowing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17169 3a8e244f-b0f2-452b-bcba-4c88e055c3ca