Adding -Wno-uninitialized to generated code
As noted by Sung, when compiling with CHPL_COMM_DEBUG set, our
generated code was triggering warnings (treated as errors for
developers) due to conservative use-before-def analysis in the C
compiler.  This change disables that warning in the generated code --
previous disablings were supported when compiling the runtime files
only, not the generated code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17174 3a8e244f-b0f2-452b-bcba-4c88e055c3ca