Fixed some futures regarding tuple variable declarations and added an optional "errorDepth" argument to compilerError and compilerWarning.
For tuple variable declarations, check that the expression on the right of the
assignment is a tuple and issue an error if it is not. Check that the size
of the tuple matches the number of variables being declared and issue an error
if not.

Added overloads of the compilerError and compilerWarning functions that take
a param int "errorDepth" as their last argument. When a depth is specified,
the error/warning is reported about a call that far up the call stack.

Added a test of the error/warning depth.
Added a semantic future about function arguments with default values after
varargs.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17181 3a8e244f-b0f2-452b-bcba-4c88e055c3ca