Make Norm use slicing rather than rank change
This is a minor change that's been in one of my working directories
since fall waiting to be tested with something more important.  This
changes the Norm module to use slicing rather than rank change, based
on a vague sense that slicing should generally be no more expensive
and might possibly be less expensive.  If anyone thinks this is
misguided, say so.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17183 3a8e244f-b0f2-452b-bcba-4c88e055c3ca